requirements.txt
twilio==8.10.0
flask==2.3.3
python-dotenv
requests
PyPDF2
langchain
langchain-community
langchain-google-genai
langchain-core
sentence-transformers
faiss-cpu
fastapi
uvicorn